+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ### Features
 
-* add InferenceTrustProfile API and controller ([5644e50](https://github.com/sonda-red/terence/commit/5644e50bee50fb3c4179d5ae37349d32d468755b))
+* add InferenceTrustBinding API and controller ([5644e50](https://github.com/sonda-red/kleym/commit/5644e50bee50fb3c4179d5ae37349d32d468755b))